Forward and back button support for mac & windows
This adds forward and back mouse button support for both macOS and
Windows.

The mac implementation is very similar to the Linux one. We keep track
of the mouse button states manually, and can re-use the same code that
we use for Linux. What differs between the platforms is what constant
Fl::event_button() returns for the respective button press. I'm not even
sure where the numbers come from and only tested on an M1 Mac.

On Windows, we call GetKeyState() directly to see if the forward and
backward mouse butons are pressed. This way, we don't have to keep track
of the button state manually.