Handle forward/back mouse button state on Windows
With this change, we keep track of forward/back (XBUTTON1/XBUTTON2)
states ourselves, just like we do with Linux and macOS. This should make
things less confusing, as we use the exact same logic for all platforms.

There are a couple of issues with this currently. The biggest one is
that you can't hold down forward/back and move the cursor outside the
window. The mouse events are captured correctly, but right now, leaving
the window will trigger an FL_LEAVE, which will send a pointerEvent to
the server with a mask of 0 which triggers a button release.

Another potential issue is that XBUTTON1 and XBUTTON2 have the values 1
and 2 respectively on WIN32, which conflict with values for normal mouse
buttons in FLTK. To circumvent this, I add an offset of 7 to whatever
the button number is, so that the result matches the button numbers in
X11 (8 and 9).