Remove msgTypePointerEventExt
Instead of adding a new message type, we can exploit the fact that the
8th bit is not used for any type of mouse events (in TigerVNC at least).

This protocol extension uses 2 bytes for extended mouse button events,
where the lower 7 bits remain the same, and the 9 upper bits are used
for the extended buttons.

The 8th bit in the mask is used to mark that the event is an extended
mouse button event, which gives us 8 more bits to use for 8 new buttons.

On the client side, upper 9 bits are shifted one step to the left, and
the 8th bit is set to 1. On the server side, the 8th bit is checked to
see if additional decoding is needed, where by the upper 8th bit is
cleared, and the upper 8 bits are shifted one step to the right.

This has to be tested with the other big VNC implementations to, since
there might be some issues with the 8th bit marker.